Move GELF chunker into its own file
